+++ src/components/App.jsx
@@ -2,19 +2,27 @@ import React from 'react';
 import { BrowserRouter as Router, Route, Switch, Redirect } from 'react-router-dom';
 import Home from '../pages/Home';
 import CreateCampaign from '../pages/CreateCampaign';
+import Login from '../pages/Login';
+import RegisterPage from '../pages/RegisterPage';
 
 export default function App() {
   return (
     <Router>
       <Switch>
-        <Route exact path="/">
+        <Route exact path='/'>
           <Home />
         </Route>
-        <Route path="/campaign/new">
+        <Route path='/campaign/new'>
           <CreateCampaign />
         </Route>
-        <Route path="*">
-          <Redirect to="/" />
+        <Route path='/login'>
+          <Login />
+        </Route>
+        <Route path='/register'>
+          <RegisterPage />
+        </Route>
+        <Route path='*'>
+          <Redirect to='/' />
         </Route>
       </Switch>
     </Router>
